requirements.txt
streamlit==1.28.0
PyPDF2
Pillow==10.0.0
langchain
langchain-huggingface
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.2.0
whisper
google-generativeai==0.5.2
gTTS==2.3.2
